requirements.txt
discord.py==2.1.0
googletrans==4.0.0-rc1
